+++ eternalcombat-plugin/src/main/java/com/eternalcode/combat/CombatPlugin.java
@@ -1,9 +1,6 @@
 package com.eternalcode.combat;
 
 import com.eternalcode.combat.bridge.BridgeService;
-import com.eternalcode.combat.command.CombatStatusCommand;
-import com.eternalcode.combat.command.CombatTagUntagCommand;
-import com.eternalcode.combat.command.CombatReloadCommand;
 import com.eternalcode.combat.command.handler.InvalidUsageHandlerImpl;
 import com.eternalcode.combat.command.handler.MissingPermissionHandlerImpl;
 import com.eternalcode.combat.config.ConfigService;
@@ -13,6 +10,7 @@ import com.eternalcode.combat.drop.DropKeepInventoryManager;
 import com.eternalcode.combat.drop.DropManager;
 import com.eternalcode.combat.drop.impl.PercentDropModifier;
 import com.eternalcode.combat.drop.impl.PlayersHealthDropModifier;
+import com.eternalcode.combat.fight.FightTagUntagCommand;
 import com.eternalcode.combat.fight.controller.FightActionBlockerController;
 import com.eternalcode.combat.fight.controller.FightMessageController;
 import com.eternalcode.combat.fight.controller.FightTagController;
@@ -29,7 +27,7 @@ import com.eternalcode.combat.fight.pearl.FightPearlController;
 import com.eternalcode.combat.fight.pearl.FightPearlManager;
 import com.eternalcode.combat.fight.tagout.FightTagOutController;
 import com.eternalcode.combat.fight.tagout.FightTagOutService;
-import com.eternalcode.combat.command.CombatTagImmunityCommand;
+import com.eternalcode.combat.fight.tagout.FightTagOutCommand;
 import com.eternalcode.combat.notification.NotificationAnnouncer;
 import com.eternalcode.combat.region.RegionController;
 import com.eternalcode.combat.region.RegionProvider;
@@ -120,10 +118,9 @@ public final class CombatPlugin extends JavaPlugin implements EternalCombatApi {
             .missingPermission(new MissingPermissionHandlerImpl(this.pluginConfig, notificationAnnouncer))
 
             .commands(
-                new CombatTagUntagCommand(this.fightManager, notificationAnnouncer, this.pluginConfig),
-                new CombatTagImmunityCommand(this.fightTagOutService, notificationAnnouncer, this.pluginConfig),
-                new CombatStatusCommand(this.fightManager, notificationAnnouncer, this.pluginConfig),
-                new CombatReloadCommand(configService, notificationAnnouncer)
+                new FightTagUntagCommand(this.fightManager, notificationAnnouncer, this.pluginConfig),
+                new FightTagOutCommand(this.fightTagOutService, notificationAnnouncer, this.pluginConfig),
+                new EternalCombatReloadCommand(configService, notificationAnnouncer)
             )
 
             .build();
